Particle system replaced by animal behavior
The system has stopped behaving like a particle system and now behaves
like organism, due to changes in constants. It is NOT behaving like
water anymore. Two forms of collision detection (if there's something in
the future position, don't go there, and  go to the average 'pushed
place' if other particles are near you) have been implemented. Need to
improve on that, as all the basic checks are for the moment n*n
(calculates the distance with all other members AND then sees if the
dist is smaller than the constant).

To implement: HashTable to improve collision-detection.